+++ test/cases/_preprocessor/ejs-require-js/src/data.js
@@ -1,0 +1,5 @@
+module.exports = {
+  title: 'Home',
+  headline: 'Breaking Bad',
+  people: ['Walter White', 'Jesse Pinkman'],
+};