+++ test/integration-pug.test.js
@@ -220,7 +220,10 @@ describe('pug-plugin tests', () => {
 
   // special cases
   test('resolve manifest.json via require', () => compareFiles('_pug/resolve-manifest.json-require'));
+
+  // TODO: fix github action issue
   test('compile template function in js', () => compareFiles('_pug/js-tmpl-entry-js'));
+
   test('inline js and css via query `?inline`', () => compareFiles('_pug/inline-js-css-query'));
 
   test('inline CSS in style tag with attributes', () => compareFiles('_pug/inline-css-in-style-tag'));